+++ DataStructureTest/InsertionSort.h
@@ -12,14 +12,13 @@ class InsertionSort : public SortTestBace
 public:
 	void Sort(std::vector<int> arr)
 	{
-		for (int i = 0; i < arr.size() - 1; i++)
+		for (size_t i = 0; i < arr.size() - 1; i++)
 		{
 			for (int j = i + 1; j > 0 && arr[j - 1] > arr[j]; j--)
 			{
 				SwapValue(arr[j], arr[j - 1]);
 			}
 		}
-
 		PrintArr(arr);
 	}
